requirements.txt
requests==2.27.1
ddt==1.4.4
xlrd==2.0.1
xlwt==1.3.0
PyYAML==6.0
